oh-tab-f2x: dedupe E2E tool-calls mock
